requirements.txt
pyserial
sounddevice
numpy